Add connection setup helpers to the quinn transport.
Copying and updating these gets really old.

This also fixes a bug in the fn make_insecure_client_endpoint that would not
support any verification schemes, causing the split example to fail.

We also add some top level type aliases and fns for creating channels for mem and testing.
