+++ Cargo.toml
@@ -31,6 +31,8 @@ postcard = { version = "1", features = ["use-std"], optional = true }
 tracing = "0.1"
 futures = { version = "0.3.30", optional = true }
 anyhow = "1.0.73"
+rcgen = { version = "0.13", optional = true }
+rustls = { version = "0.23", default-features = false, features = ["ring"], optional = true }
 
 # Indirect dependencies, is needed to make the minimal crates versions work
 slab = "0.4.9" # iroh-quinn
@@ -60,6 +62,7 @@ quinn-transport = ["dep:flume", "dep:quinn", "dep:postcard", "dep:bytes", "dep:t
 flume-transport = ["dep:flume"]
 iroh-transport = ["dep:iroh", "dep:flume", "dep:postcard", "dep:tokio-serde", "tokio-util/codec"]
 macros = []
+test-utils = ["dep:rcgen", "dep:rustls"]
 default = ["flume-transport"]
 
 [package.metadata.docs.rs]
