+++ examples/split/client/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0.14"
 futures = "0.3.26"
-quic-rpc = { path = "../../..", features = ["quinn-transport", "macros"] }
+quic-rpc = { path = "../../..", features = ["quinn-transport", "macros", "test-utils"] }
 quinn = { package = "iroh-quinn", version = "0.12" }
 rustls = { version = "0.23", default-features = false, features = ["ring"] }
 tracing-subscriber = "0.3.16"
