+++ src/transport/quinn.rs
@@ -803,3 +803,146 @@ pub fn get_handshake_data(
         server_name: tls_connection.server_name.clone(),
     })
 }
+
+#[cfg(feature = "test-utils")]
+mod quinn_setup_utils {
+    use std::{net::SocketAddr, sync::Arc};
+
+    use anyhow::Result;
+    use quinn::{crypto::rustls::QuicClientConfig, ClientConfig, Endpoint, ServerConfig};
+
+    /// Builds default quinn client config and trusts given certificates.
+    ///
+    /// ## Args
+    ///
+    /// - server_certs: a list of trusted certificates in DER format.
+    pub fn configure_client(server_certs: &[&[u8]]) -> Result<ClientConfig> {
+        let mut certs = rustls::RootCertStore::empty();
+        for cert in server_certs {
+            let cert = rustls::pki_types::CertificateDer::from(cert.to_vec());
+            certs.add(cert)?;
+        }
+
+        let crypto_client_config = rustls::ClientConfig::builder_with_provider(Arc::new(
+            rustls::crypto::ring::default_provider(),
+        ))
+        .with_protocol_versions(&[&rustls::version::TLS13])
+        .expect("valid versions")
+        .with_root_certificates(certs)
+        .with_no_client_auth();
+        let quic_client_config =
+            quinn::crypto::rustls::QuicClientConfig::try_from(crypto_client_config)?;
+
+        Ok(ClientConfig::new(Arc::new(quic_client_config)))
+    }
+
+    /// Constructs a QUIC endpoint configured for use a client only.
+    ///
+    /// ## Args
+    ///
+    /// - server_certs: list of trusted certificates.
+    pub fn make_client_endpoint(bind_addr: SocketAddr, server_certs: &[&[u8]]) -> Result<Endpoint> {
+        let client_cfg = configure_client(server_certs)?;
+        let mut endpoint = Endpoint::client(bind_addr)?;
+        endpoint.set_default_client_config(client_cfg);
+        Ok(endpoint)
+    }
+
+    /// Create a server endpoint with a self-signed certificate
+    ///
+    /// Returns the server endpoint and the certificate in DER format
+    pub fn make_server_endpoint(bind_addr: SocketAddr) -> Result<(Endpoint, Vec<u8>)> {
+        let (server_config, server_cert) = configure_server()?;
+        let endpoint = Endpoint::server(server_config, bind_addr)?;
+        Ok((endpoint, server_cert))
+    }
+
+    /// Create a quinn server config with a self-signed certificate
+    ///
+    /// Returns the server config and the certificate in DER format
+    pub fn configure_server() -> anyhow::Result<(ServerConfig, Vec<u8>)> {
+        let cert = rcgen::generate_simple_self_signed(vec!["localhost".into()])?;
+        let cert_der = cert.cert.der();
+        let priv_key = rustls::pki_types::PrivatePkcs8KeyDer::from(cert.key_pair.serialize_der());
+        let cert_chain = vec![cert_der.clone()];
+
+        let mut server_config = ServerConfig::with_single_cert(cert_chain, priv_key.into())?;
+        Arc::get_mut(&mut server_config.transport)
+            .unwrap()
+            .max_concurrent_uni_streams(0_u8.into());
+
+        Ok((server_config, cert_der.to_vec()))
+    }
+
+    /// Constructs a QUIC endpoint that trusts all certificates.
+    ///
+    /// This is useful for testing and local connections, but should be used with care.
+    pub fn make_insecure_client_endpoint(bind_addr: SocketAddr) -> Result<Endpoint> {
+        let crypto = rustls::ClientConfig::builder()
+            .dangerous()
+            .with_custom_certificate_verifier(Arc::new(SkipServerVerification))
+            .with_no_client_auth();
+
+        let client_cfg = QuicClientConfig::try_from(crypto)?;
+        let client_cfg = ClientConfig::new(Arc::new(client_cfg));
+        let mut endpoint = Endpoint::client(bind_addr)?;
+        endpoint.set_default_client_config(client_cfg);
+        Ok(endpoint)
+    }
+
+    #[derive(Debug)]
+    struct SkipServerVerification;
+
+    impl rustls::client::danger::ServerCertVerifier for SkipServerVerification {
+        fn verify_server_cert(
+            &self,
+            _end_entity: &rustls::pki_types::CertificateDer<'_>,
+            _intermediates: &[rustls::pki_types::CertificateDer<'_>],
+            _server_name: &rustls::pki_types::ServerName<'_>,
+            _ocsp_response: &[u8],
+            _now: rustls::pki_types::UnixTime,
+        ) -> Result<rustls::client::danger::ServerCertVerified, rustls::Error> {
+            Ok(rustls::client::danger::ServerCertVerified::assertion())
+        }
+
+        fn verify_tls12_signature(
+            &self,
+            _message: &[u8],
+            _cert: &rustls::pki_types::CertificateDer<'_>,
+            _dss: &rustls::DigitallySignedStruct,
+        ) -> Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
+            Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
+        }
+
+        fn verify_tls13_signature(
+            &self,
+            _message: &[u8],
+            _cert: &rustls::pki_types::CertificateDer<'_>,
+            _dss: &rustls::DigitallySignedStruct,
+        ) -> Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
+            Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
+        }
+
+        fn supported_verify_schemes(&self) -> Vec<rustls::SignatureScheme> {
+            use rustls::SignatureScheme::*;
+            // list them all, we don't care.
+            vec![
+                RSA_PKCS1_SHA1,
+                ECDSA_SHA1_Legacy,
+                RSA_PKCS1_SHA256,
+                ECDSA_NISTP256_SHA256,
+                RSA_PKCS1_SHA384,
+                ECDSA_NISTP384_SHA384,
+                RSA_PKCS1_SHA512,
+                ECDSA_NISTP521_SHA512,
+                RSA_PSS_SHA256,
+                RSA_PSS_SHA384,
+                RSA_PSS_SHA512,
+                ED25519,
+                ED448,
+            ]
+        }
+    }
+}
+#[cfg(feature = "test-utils")]
+pub use quinn_setup_utils::*;
